refactor: stabilize token detail chart and secure order detail flow
