--- client/web/src/pages/MockupPage.jsx
@@ -1,2299 +1,0 @@
-﻿// MockupPage — 상세 페이지 실데이터 연동 목업
-// /token/:tokenId 로 접근
-// 팀원 작업 중인 TradingPage와 분리된 독립 테스트 공간
-
-import { useState, useEffect, useCallback, useMemo } from 'react';
-import { useParams } from 'react-router-dom';
-import {
-  ResponsiveContainer, ComposedChart, BarChart, Bar, Cell,
-  XAxis, YAxis, Tooltip, CartesianGrid,
-} from 'recharts';
-import {
-  Maximize2, X, Target, MoreHorizontal, Edit3,
-} from 'lucide-react';
-
-import { useApp }            from '../context/AppContext.jsx';
-import { useTradingSocket }  from '../hooks/useTradingSocket.js';
-import { AssetHeader }       from '../components/trading/AssetHeader.jsx';
-import { SecureOrderPanel }  from '../components/trading/SecureOrderPanel.jsx';
-import { HogaRow }           from '../components/trading/HogaRow.jsx';
-import { PriceRow }          from '../components/trading/PriceRow.jsx';
-import { cn } from '../lib/utils.js';
-import { API_BASE_URL } from '../lib/config.js';
-import api from '../lib/api.js';
-
-const API = API_BASE_URL;
-
-// JWT payload에서 memberId(sub 클레임) 추출
-function parseJwtMemberId(token) {
-  if (!token) return null;
-  try {
-    const payloadPart = token.split('.')[1];
-    if (!payloadPart) return null;
-    const base64 = payloadPart
-        .replace(/-/g, '+')
-        .replace(/_/g, '/')
-        .padEnd(Math.ceil(payloadPart.length / 4) * 4, '=');
-    const payload = JSON.parse(atob(base64));
-    const memberId = Number(payload?.sub);
-    return Number.isFinite(memberId) && Number.isInteger(memberId) ? memberId : null;
-  } catch {
-    return null;
-  }
-}
-
-const PERIOD_TO_TYPE = {
-  '분':  'MINUTE',
-  '시간': 'HOUR',
-  '일':  'DAY',
-  '월':  'MONTH',
-  '년':  'YEAR',
-};
-const CHART_PERIODS = ['분', '시간', '일', '월', '년'];
-
-// ── 유틸 ────────────────────────────────────────────────────────
-function formatCandleTime(candleTime, period) {
-  if (!candleTime) return '';
-  const d = new Date(candleTime);
-  if (period === '분' || period === '시간') return d.toTimeString().slice(0, 5);
-  if (period === '일') return `${d.getMonth() + 1}/${d.getDate()}`;
-  if (period === '월') return `${d.getFullYear()}/${String(d.getMonth() + 1).padStart(2, '0')}`;
-  if (period === '년') return `${d.getFullYear()}`;
-  return d.toTimeString().slice(0, 5);
-}
-
-// WebSocket tradeTime 포맷 (Jackson 배열 [y,mo,d,h,m,s] 또는 ISO 문자열 모두 처리)
-function formatTradeTime(t) {
-  if (!t) return '';
-  if (Array.isArray(t)) {
-    const [, , , h = 0, m = 0, s = 0] = t;
-    return `${String(h).padStart(2, '0')}:${String(m).padStart(2, '0')}:${String(s).padStart(2, '0')}`;
-  }
-  try {
-    return new Date(t).toLocaleTimeString('ko-KR', { hour: '2-digit', minute: '2-digit', second: '2-digit', hour12: false });
-  } catch { return String(t); }
-}
-
-function mapCandle(dto, period) {
-  return {
-    ts:    dto.candleTime ? new Date(dto.candleTime).getTime() : 0,
-    time:  formatCandleTime(dto.candleTime, period),
-    open:  Math.round(dto.openPrice  || 0),
-    high:  Math.round(dto.highPrice  || 0),
-    low:   Math.round(dto.lowPrice   || 0),
-    close: Math.round(dto.closePrice || 0),
-    vol:   Math.round(dto.volume     || 0),
-  };
-}
-
-// 현재 시간 기준 35개 고정 슬롯 생성 (오래된 → 최근 순, 현재 분 제외)
-function generate35Slots(period) {
-  const now = new Date();
-  const slots = [];
-  for (let i = 34; i >= 0; i--) {
-    const d = new Date(now);
-    if (period === '분') {
-      d.setSeconds(0, 0);
-      d.setMinutes(d.getMinutes() - (i + 1));
-    } else if (period === '시간') {
-      d.setMinutes(0, 0, 0);
-      d.setHours(d.getHours() - (i + 1));
-    } else if (period === '일') {
-      d.setHours(0, 0, 0, 0);
-      d.setDate(d.getDate() - (i + 1));
-    } else if (period === '월') {
-      d.setDate(1); d.setHours(0, 0, 0, 0);
-      d.setMonth(d.getMonth() - (i + 1));
-    } else if (period === '년') {
-      d.setMonth(0, 1); d.setHours(0, 0, 0, 0);
-      d.setFullYear(d.getFullYear() - (i + 1));
-    }
-    slots.push({
-      ts: d.getTime(),
-      time: formatCandleTime(d.getTime(), period),
-      open: null, high: null, low: null, close: null, vol: null,
-    });
-  }
-  return slots;
-}
-
-// 35개 슬롯 + 실제 캔들 병합 — 데이터 없는 슬롯은 null OHLCV 유지
-function buildChartData(fetchedCandles, period) {
-  const slots = generate35Slots(period);
-  const candleByTime = new Map(fetchedCandles.map(c => [c.time, c]));
-  return slots.map(slot => candleByTime.get(slot.time) ?? slot);
-}
-
-// ── 캔들스틱 shape ───────────────────────────────────────────────
-// dataKey={d => [d.low, d.high]} 와 함께 사용
-// Recharts: y = high 의 픽셀 위치(위쪽), y+height = low 의 픽셀 위치(아래쪽)
-// → Y축 auto domain이 high/low 전체를 포함하므로 wick 이 잘리지 않음
-function CandlestickShape(props) {
-  const { x, y, width, height } = props;
-  const open  = props.payload?.open;
-  const close = props.payload?.close;
-  const high  = props.payload?.high;
-  const low   = props.payload?.low;
-
-  if (open == null || close == null || high == null || low == null) return null;
-  if (width <= 0 || height <= 0) return null;
-
-  const priceRange = high - low;
-  if (priceRange <= 0) return null;
-
-  const isUp  = close >= open;
-  const color = isUp ? '#e54d4d' : '#3b82f6';
-  const cx    = x + width / 2;
-  const ratio = height / priceRange;  // px per price unit
-
-  // y      = high 픽셀 (차트 위쪽)
-  // y+height = low 픽셀 (차트 아래쪽)
-  const highPx = y;
-  const lowPx  = y + height;
-
-  const bodyTopPx    = y + (high - Math.max(open, close)) * ratio;
-  const bodyBottomPx = y + (high - Math.min(open, close)) * ratio;
-  const bodyH        = Math.max(bodyBottomPx - bodyTopPx, 1);
-
-  return (
-      <g>
-        {/* 위 꼬리: high → body 상단 */}
-        <line x1={cx} y1={highPx} x2={cx} y2={bodyTopPx}
-              stroke={color} strokeWidth={1.2} />
-        {/* 아래 꼬리: body 하단 → low */}
-        <line x1={cx} y1={bodyBottomPx} x2={cx} y2={lowPx}
-              stroke={color} strokeWidth={1.2} />
-        {/* 몸통 */}
-        <rect x={x + 1} y={bodyTopPx} width={Math.max(width - 2, 1)} height={bodyH}
-              fill={color} rx={1} />
-      </g>
-  );
-}
-
-// ── 로그인 필요 팝업 ────────────────────────────────────────────
-function LoginRequiredModal({ message, onClose }) {
-  return (
-      <div className="fixed inset-0 z-50 flex items-center justify-center bg-black/40"
-           onClick={onClose}>
-        <div className="bg-white rounded-2xl border border-stone-200 shadow-xl p-8 w-80 text-center"
-             onClick={e => e.stopPropagation()}>
-          <p className="text-sm font-bold text-stone-800 mb-6">{message}</p>
-          <button
-              onClick={onClose}
-              className="w-full py-3 bg-stone-800 text-white text-sm font-black rounded-xl hover:bg-stone-700 transition-colors"
-          >
-            확인
-          </button>
-        </div>
-
-        {false && passwordModal && (
-            <OrderPinPadModal
-                title={
-                  passwordModal.action === 'create'
-                      ? `${isBuy ? '매수' : '매도'} 주문 확인`
-                      : passwordModal.action === 'update'
-                          ? '주문 수정 확인'
-                          : '주문 취소 확인'
-                }
-                description={
-                  passwordModal.action === 'create'
-                      ? `${isBuy ? '매수' : '매도'} 주문 내역을 확인한 뒤 계좌 비밀번호를 입력해 주세요.`
-                      : passwordModal.action === 'update'
-                          ? '수정할 주문 내역을 확인한 뒤 계좌 비밀번호를 입력해 주세요.'
-                          : '취소할 주문 내역을 확인한 뒤 계좌 비밀번호를 입력해 주세요.'
-                }
-                password={passwordModal.action === 'create' ? accountPassword : editAccountPassword}
-                errorMessage={
-                  passwordModal.action === 'create'
-                      ? (orderMsg?.type === 'error' ? orderMsg.text : null)
-                      : (updateMsg?.orderId === passwordModal.orderId && updateMsg.type === 'error' ? updateMsg.text : null)
-                }
-                submitting={submitting}
-                onChange={value => {
-                  if (passwordModal.action === 'create') setAccountPassword(value);
-                  else setEditAccountPassword(value);
-                }}
-                onClose={closePasswordModal}
-                onConfirm={handlePasswordConfirm}
-            />
-        )}
-      </div>
-  );
-}
-
-function OrderPinPadModal({
-  title,
-  description,
-  password,
-  errorMessage,
-  submitting,
-  onChange,
-  onClose,
-  onConfirm,
-}) {
-  const masked = password.length > 0 ? '•'.repeat(password.length) : '○ ○ ○ ○';
-  const keys = ['1', '2', '3', '4', '5', '6', '7', '8', '9', 'reset', '0', 'delete'];
-
-  return (
-      <div className="fixed inset-0 z-[60] flex items-center justify-center bg-black/50" onClick={onClose}>
-        <div
-            className="w-[360px] rounded-2xl border border-stone-200 bg-white p-6 shadow-xl"
-            onClick={e => e.stopPropagation()}
-        >
-          <div className="space-y-2">
-            <h3 className="text-base font-black text-stone-800">{title}</h3>
-            <p className="text-sm font-medium text-stone-500">{description}</p>
-          </div>
-
-          <div className="mt-5 rounded-2xl border border-stone-200 bg-stone-100 px-4 py-5">
-            <div className="text-center font-mono text-2xl font-black tracking-[0.35em] text-stone-800">
-              {masked}
-            </div>
-          </div>
-
-          {errorMessage && (
-              <p className="mt-3 text-center text-[11px] font-bold text-brand-red">{errorMessage}</p>
-          )}
-
-          <div className="mt-5 grid grid-cols-3 gap-2">
-            {keys.map(key => (
-                <button
-                    key={key}
-                    type="button"
-                    onClick={() => {
-                      if (key === 'reset') {
-                        onChange('');
-                        return;
-                      }
-                      if (key === 'delete') {
-                        onChange(password.slice(0, -1));
-                        return;
-                      }
-                      if (password.length >= 4) return;
-                      onChange(`${password}${key}`);
-                    }}
-                    className={cn(
-                        'rounded-xl border py-3 text-sm font-black transition-colors',
-                        key === 'reset' || key === 'delete'
-                            ? 'border-stone-200 bg-stone-100 text-stone-500 hover:bg-stone-200'
-                            : 'border-stone-200 bg-white text-stone-800 hover:bg-stone-100'
-                    )}
-                >
-                  {key === 'reset' ? '초기화' : key === 'delete' ? '지우기' : key}
-                </button>
-            ))}
-          </div>
-
-          <div className="mt-6 rounded-2xl border border-stone-200 bg-stone-50 p-4 text-sm text-stone-600">
-            {description}
-          </div>
-
-          <div className="mt-4 flex gap-2">
-            <button
-                onClick={onClose}
-                disabled={submitting}
-                className="flex-1 rounded-xl border border-stone-200 bg-white py-3 text-sm font-black text-stone-500 hover:bg-stone-100 disabled:opacity-50"
-            >
-              취소
-            </button>
-            <button
-                onClick={onConfirm}
-                disabled={submitting || password.length !== 4}
-                className="flex-1 rounded-xl bg-stone-800 py-3 text-sm font-black text-white hover:bg-stone-700 disabled:opacity-50"
-            >
-              {submitting ? '처리 중...' : '확인'}
-            </button>
-          </div>
-        </div>
-
-        {false && passwordModal && (
-            <OrderPinPadModal
-                title={
-                  passwordModal.action === 'create'
-                      ? `${isBuy ? '매수' : '매도'} 주문 확인`
-                      : passwordModal.action === 'update'
-                          ? '주문 수정 확인'
-                          : '주문 취소 확인'
-                }
-                description={
-                  passwordModal.action === 'create'
-                      ? `${isBuy ? '매수' : '매도'} 주문 내역을 확인한 뒤 계좌 비밀번호를 입력해 주세요.`
-                      : passwordModal.action === 'update'
-                          ? '수정할 주문 내역을 확인한 뒤 계좌 비밀번호를 입력해 주세요.'
-                          : '취소할 주문 내역을 확인한 뒤 계좌 비밀번호를 입력해 주세요.'
-                }
-                password={passwordModal.action === 'create' ? accountPassword : editAccountPassword}
-                errorMessage={
-                  passwordModal.action === 'create'
-                      ? (orderMsg?.type === 'error' ? orderMsg.text : null)
-                      : (updateMsg?.orderId === passwordModal.orderId && updateMsg.type === 'error' ? updateMsg.text : null)
-                }
-                submitting={submitting}
-                onChange={value => {
-                  if (passwordModal.action === 'create') setAccountPassword(value);
-                  else setEditAccountPassword(value);
-                }}
-                onClose={closePasswordModal}
-                onConfirm={handlePasswordConfirm}
-            />
-        )}
-      </div>
-  );
-}
-
-function OrderExecutionConfirmModal({
-  title,
-  items,
-  submitting,
-  onClose,
-  onConfirm,
-}) {
-  return (
-      <div className="fixed inset-0 z-[70] flex items-center justify-center bg-black/55" onClick={onClose}>
-        <div
-            className="w-[360px] rounded-2xl border border-stone-200 bg-white p-6 shadow-xl"
-            onClick={e => e.stopPropagation()}
-        >
-          <div className="space-y-1">
-            <h3 className="text-base font-black text-stone-800">{title}</h3>
-            <p className="text-sm text-stone-500">주문 내역을 확인한 뒤 최종 진행해 주세요.</p>
-          </div>
-
-          <div className="mt-5 space-y-2 rounded-2xl border border-stone-200 bg-stone-50 p-4">
-            {items.map(item => (
-                <div key={item.label} className="flex items-center justify-between text-sm">
-                  <span className="font-bold text-stone-400">{item.label}</span>
-                  <span className="font-mono font-black text-stone-800">{item.value}</span>
-                </div>
-            ))}
-          </div>
-
-          <div className="mt-6 flex gap-2">
-            <button
-                onClick={onClose}
-                disabled={submitting}
-                className="flex-1 rounded-xl border border-stone-200 bg-white py-3 text-sm font-black text-stone-500 hover:bg-stone-100 disabled:opacity-50"
-            >
-              취소
-            </button>
-            <button
-                onClick={onConfirm}
-                disabled={submitting}
-                className="flex-1 rounded-xl bg-stone-800 py-3 text-sm font-black text-white hover:bg-stone-700 disabled:opacity-50"
-            >
-              {submitting ? '처리 중...' : '최종 확인'}
-            </button>
-          </div>
-        </div>
-      </div>
-  );
-}
-
-function OrderPasswordModal({
-  title,
-  description,
-  password,
-  errorMessage,
-  submitting,
-  onChange,
-  onClose,
-  onConfirm,
-}) {
-  return (
-      <div className="fixed inset-0 z-50 flex items-center justify-center bg-black/40" onClick={onClose}>
-        <div
-            className="w-[360px] rounded-2xl border border-stone-200 bg-white p-6 shadow-xl"
-            onClick={e => e.stopPropagation()}
-        >
-          <div className="space-y-2">
-            <h3 className="text-base font-black text-stone-800">{title}</h3>
-            <p className="text-sm font-medium text-stone-500">{description}</p>
-          </div>
-
-          <div className="mt-5 space-y-2">
-            <label className="text-[11px] font-bold text-stone-400">계좌 비밀번호</label>
-            <div className="flex items-center gap-2 rounded-xl border border-stone-200 bg-stone-100 px-4 py-3">
-              <input
-                  type="password"
-                  inputMode="numeric"
-                  maxLength={4}
-                  autoFocus
-                  value={password}
-                  onChange={e => onChange(e.target.value.replace(/\D/g, '').slice(0, 4))}
-                  placeholder="숫자 4자리"
-                  className="flex-1 bg-transparent text-right font-mono text-sm font-bold text-stone-800 outline-none placeholder-stone-400"
-              />
-            </div>
-            {errorMessage && <p className="text-[11px] font-bold text-brand-red">{errorMessage}</p>}
-          </div>
-
-          <div className="mt-6 flex gap-2">
-            <button
-                onClick={onClose}
-                disabled={submitting}
-                className="flex-1 rounded-xl border border-stone-200 bg-white py-3 text-sm font-black text-stone-500 transition-colors hover:bg-stone-100 disabled:opacity-50"
-            >
-              취소
-            </button>
-            <button
-                onClick={onConfirm}
-                disabled={submitting}
-                className="flex-1 rounded-xl bg-stone-800 py-3 text-sm font-black text-white transition-colors hover:bg-stone-700 disabled:opacity-50"
-            >
-              {submitting ? '처리 중...' : '확인'}
-            </button>
-          </div>
-        </div>
-
-        {false && passwordModal && (
-            <OrderPasswordModal
-                title={
-                  passwordModal.action === 'create'
-                      ? `${isBuy ? '매수' : '매도'} 주문 확인`
-                      : passwordModal.action === 'update'
-                          ? '주문 수정 확인'
-                          : '주문 취소 확인'
-                }
-                description={
-                  passwordModal.action === 'create'
-                      ? '계좌 비밀번호를 입력한 뒤 주문을 진행해 주세요.'
-                      : passwordModal.action === 'update'
-                          ? '계좌 비밀번호를 입력한 뒤 주문 수정을 진행해 주세요.'
-                          : '계좌 비밀번호를 입력한 뒤 주문 취소를 진행해 주세요.'
-                }
-                password={passwordModal.action === 'create' ? accountPassword : editAccountPassword}
-                errorMessage={
-                  passwordModal.action === 'create'
-                      ? (orderMsg?.type === 'error' ? orderMsg.text : null)
-                      : (updateMsg?.orderId === passwordModal.orderId && updateMsg.type === 'error' ? updateMsg.text : null)
-                }
-                submitting={submitting}
-                onChange={value => {
-                  if (passwordModal.action === 'create') setAccountPassword(value);
-                  else setEditAccountPassword(value);
-                }}
-                onClose={closePasswordModal}
-                onConfirm={handlePasswordConfirm}
-            />
-        )}
-      </div>
-  );
-}
-
-// ── MockupPage ───────────────────────────────────────────────────
-export function MockupPage() {
-  const { tokenId }    = useParams();
-  const { user, likedTokenIds, toggleLike } = useApp();
-  const TOKEN_ID = Number(tokenId) || 1;
-  const memberId = parseJwtMemberId(user?.accessToken);
-
-  // ── 로그인 필요 팝업 ─────────────────────────────────────────
-  const [loginModal, setLoginModal] = useState(null); // null | string(message)
-
-  // ── 탭 상태 ─────────────────────────────────────────────────
-  const [activeTab, setActiveTab] = useState('chart');
-
-  // ── 토큰 상세 정보 (백엔드) ──────────────────────────────────
-  const [tokenInfo, setTokenInfo] = useState(null);
-
-  useEffect(() => {
-    api.get(`/api/token/${TOKEN_ID}/chart`)
-        .then(r => setTokenInfo(r.data))
-        .catch(e => console.warn('[MockupPage] 토큰 상세 조회 실패:', e));
-  }, [TOKEN_ID, user?.accessToken]);
-
-  // ── 차트 상태 ────────────────────────────────────────────────
-  const [chartPeriod, setChartPeriod]   = useState('분');
-  const [chartData, setChartData]       = useState([]);
-  const [hoveredData, setHoveredData]   = useState(null);
-  const [loading, setLoading]           = useState(false);
-  const [chartModalOpen, setChartModalOpen] = useState(false);
-
-  // ── 종목 정보 탭 데이터 ─────────────────────────────────────
-  const [tokenAssetInfo, setTokenAssetInfo] = useState(null);
-
-  useEffect(() => {
-    if (activeTab !== 'info' || tokenAssetInfo !== null) return;
-    api.get(`/api/token/${TOKEN_ID}/info`)
-        .then(r => setTokenAssetInfo(r.data))
-        .catch(e => { console.warn('[MockupPage] 종목정보 조회 실패:', e); setTokenAssetInfo({}); });
-  }, [activeTab, TOKEN_ID, user?.accessToken]);
-
-  // ── 배당금 / 공시 탭 데이터 ──────────────────────────────────
-  const [allocations, setAllocations]   = useState(null);
-  const [disclosures, setDisclosures]   = useState(null);
-
-  useEffect(() => {
-    if (activeTab !== 'dividend' || allocations !== null) return;
-    api.get(`/api/token/${TOKEN_ID}/allocation`)
-        .then(r => setAllocations(r.data))
-        .catch(e => { console.warn('[MockupPage] 배당금 조회 실패:', e); setAllocations([]); });
-  }, [activeTab, TOKEN_ID, user?.accessToken]);
-
-  useEffect(() => {
-    if (activeTab !== 'news' || disclosures !== null) return;
-    api.get(`/api/token/${TOKEN_ID}/disclosure`)
-        .then(r => setDisclosures(r.data))
-        .catch(e => { console.warn('[MockupPage] 공시 조회 실패:', e); setDisclosures([]); });
-  }, [activeTab, TOKEN_ID, user?.accessToken]);
-
-  // ── 호가 / 체결 상태 ────────────────────────────────────────
-  // 호가: WebSocket snapshot이 구독 즉시 전송하므로 빈 배열로 초기화
-  const [asks, setAsks]             = useState([]);
-  const [bids, setBids]             = useState([]);
-  const [executions, setExecutions] = useState([]);
-  const [trades, setTrades]         = useState([]);
-
-  // ── 오늘 최고/최저가 — REST 초기값 + DAY 캔들 WebSocket 실시간 갱신 ──
-  const [todayHigh, setTodayHigh] = useState(null);
-  const [todayLow,  setTodayLow]  = useState(null);
-
-  // tokenInfo 로드 완료 시 초기값 세팅 (오늘 체결이 없으면 null 유지)
-  useEffect(() => {
-    if (tokenInfo?.todayHighPrice) setTodayHigh(tokenInfo.todayHighPrice);
-    if (tokenInfo?.todayLowPrice)  setTodayLow(tokenInfo.todayLowPrice);
-  }, [tokenInfo]);
-
-  // ── 체결 목록 초기 REST 로드 ─────────────────────────────────
-  useEffect(() => {
-    api.get(`/api/token/${TOKEN_ID}/trades`)
-        .then(r => {
-          setTrades(r.data.map(d => ({
-            price:      d.tradePrice,
-            qty:        d.tradeQuantity,
-            changeRate: d.percentageChange ?? 0,
-            vol:        d.totalVolume ?? 0,
-            time:       d.executedAt
-                ? new Date(d.executedAt).toLocaleTimeString('ko-KR', { hour: '2-digit', minute: '2-digit', second: '2-digit', hour12: false })
-                : '',
-          })));
-        })
-        .catch(e => console.warn('[MockupPage] 체결 목록 조회 실패:', e));
-  }, [TOKEN_ID, user?.accessToken]);
-
-  // ── 현재가 — REST 초기값 + 체결 WebSocket으로만 갱신 (차트 주기와 무관)
-  const [currentPrice, setCurrentPrice] = useState(0);
-
-  useEffect(() => {
-    if (tokenInfo?.currentPrice) setCurrentPrice(tokenInfo.currentPrice);
-  }, [tokenInfo]);
-
-  // 차트 hover용 — 차트 주기 전환에 영향받지 않도록 currentPrice와 분리
-  const lastWithData = useMemo(
-    () => [...chartData].reverse().find(c => c.close != null),
-    [chartData]
-  );
-
-  const basePrice = tokenInfo?.yesterdayClosePrice || currentPrice || 1;
-  const displayData = hoveredData || lastWithData || null;
-
-  const validHighs = useMemo(() => chartData.map(c => c.high).filter(v => v != null), [chartData]);
-  const validLows  = useMemo(() => chartData.map(c => c.low).filter(v  => v != null), [chartData]);
-  const dailyHigh  = validHighs.length > 0 ? Math.max(...validHighs) : 0;
-  const dailyLow   = validLows.length  > 0 ? Math.min(...validLows)  : 0;
-
-  // Y축 도메인 — null 슬롯 제외한 실제 가격 기준
-  // 호가 단위 정책 (TickSizePolicy.java 동일 기준)
-  const getTickSize = (price) => {
-    if (price < 100)   return 10;
-    if (price < 1000)  return 50;
-    if (price < 10000) return 100;
-    return 500;
-  };
-
-  const { yDomain, yTicks } = useMemo(() => {
-    const prices = chartData.flatMap(d => [d.low, d.high]).filter(v => v != null);
-    if (prices.length === 0) return { yDomain: ['auto', 'auto'], yTicks: undefined };
-
-    const min = Math.min(...prices);
-    const max = Math.max(...prices);
-    const mid = (min + max) / 2;
-    const tickSize = getTickSize(mid);
-
-    // 패딩: 최소 tickSize 1칸, 범위의 10%
-    const pad = Math.max(tickSize, Math.ceil((max - min) * 0.1 / tickSize) * tickSize);
-    const domainMin = Math.floor((min - pad) / tickSize) * tickSize;
-    const domainMax = Math.ceil((max + pad) / tickSize) * tickSize;
-
-    // 5~7개 tick — 호가 단위 배수로 정렬
-    const range = domainMax - domainMin;
-    const interval = Math.ceil(range / (5 * tickSize)) * tickSize;
-    const ticks = [];
-    for (let t = domainMin; t <= domainMax; t += interval) ticks.push(t);
-
-    return { yDomain: [domainMin, domainMax], yTicks: ticks };
-  }, [chartData]);
-
-  const maxAskAmount = Math.max(...asks.map(r => r.amount), 1);
-  const maxBidAmount = Math.max(...bids.map(r => r.amount), 1);
-
-  // AssetHeader 가 기대하는 asset 객체 형태로 변환
-  // tokenInfo 로드 전에도 탭이 보여야 하므로 항상 객체를 반환
-  const changeRate = basePrice > 0 && currentPrice > 0
-      ? Math.round(((currentPrice - basePrice) / basePrice) * 10000) / 100
-      : 0;
-
-  const asset = {
-    id:     TOKEN_ID,
-    name:   tokenInfo?.tokenName  || tokenInfo?.assetName || '-',
-    symbol: tokenInfo?.tokenSymbol || '-',
-    change: changeRate,
-    high:   dailyHigh || currentPrice || 0,
-    low:    dailyLow  || currentPrice || 0,
-    price:  currentPrice,
-    issued: tokenInfo?.totalSupply || 0,
-    desc:   tokenInfo?.assetName   || '',
-    pdfUrl: null,
-  };
-
-  const isLiked = likedTokenIds.includes(TOKEN_ID);
-
-  // ── 캔들 API 조회 ────────────────────────────────────────────
-  const fetchCandles = useCallback(async () => {
-    setLoading(true);
-    try {
-      const type = PERIOD_TO_TYPE[chartPeriod];
-      const res = await api.get(`/api/token/${TOKEN_ID}/candle?type=${type}`);
-      setChartData(buildChartData(res.data.map(d => mapCandle(d, chartPeriod)), chartPeriod));
-    } catch (e) {
-      console.warn('[MockupPage] 캔들 조회 실패:', e.message);
-    } finally {
-      setLoading(false);
-    }
-  }, [chartPeriod, TOKEN_ID, user?.accessToken]);
-
-  useEffect(() => {
-    setChartData([]);   // 기간 변경 시 이전 기간 데이터 초기화 (분/시/일 혼재 방지)
-    fetchCandles();
-  }, [fetchCandles]);
-
-  // ── 대기 주문 WS 업데이트 ────────────────────────────────────
-  // match 서버가 pendingOrders:{tokenId}:{memberId} publish 시 LoginGateOrderPanel로 전달
-  const [wsPendingData, setWsPendingData] = useState(null);
-
-  // ── WebSocket 연동 ───────────────────────────────────────────
-  // /topic/candle/live/{tokenId}/{candleType} 구독
-  // 현재 선택된 주기(chartPeriod)의 타입만 구독 → 차트 전환 시 재연결
-  useTradingSocket({
-    tokenId:    TOKEN_ID,
-    candleType: PERIOD_TO_TYPE[chartPeriod],
-    token:      user?.accessToken,
-    memberId,
-    onOrderBook: (data) => {
-      if (data.asks) setAsks(data.asks.map(r => ({ price: r.price, amount: r.quantity })));
-      if (data.bids) setBids(data.bids.map(r => ({ price: r.price, amount: r.quantity })));
-    },
-    onTrades: (data) => {
-      if (data.tradePrice) setCurrentPrice(data.tradePrice);
-      setExecutions(prev =>
-          [{ price: data.tradePrice, qty: data.tradeQuantity, isBuy: data.isBuy }, ...prev].slice(0, 15)
-      );
-      setTrades(prev => {
-        const cumulativeVol = (prev[0]?.vol ?? 0) + (data.tradeQuantity ?? 0);
-        // 전날 종가 대비 등락률 — REST API와 동일한 기준
-        const base = tokenInfo?.yesterdayClosePrice;
-        const rate = base > 0 ? ((data.tradePrice - base) / base) * 100 : 0;
-        return [{
-          price:      data.tradePrice,
-          qty:        data.tradeQuantity,
-          changeRate: Math.round(rate * 100) / 100,
-          vol:        cumulativeVol,
-          time:       formatTradeTime(data.tradeTime),
-        }, ...prev].slice(0, 50);
-      });
-    },
-    onCandle: (data) => {
-      const newCandle = mapCandle(data, chartPeriod);
-      setChartData(prev => {
-        if (prev.length === 0) return [newCandle];
-        const idx = prev.findIndex(c => c.time === newCandle.time);
-        if (idx >= 0) {
-          // 기존 슬롯(빈 슬롯 포함) 교체
-          const updated = [...prev];
-          updated[idx] = newCandle;
-          return updated;
-        }
-        // 새 시간대 캔들 — 오른쪽 끝에 추가
-        return [...prev, newCandle];
-      });
-    },
-    onDayCandle: (data) => {
-      if (data.highPrice != null) setTodayHigh(h => h == null ? data.highPrice : Math.max(h, data.highPrice));
-      if (data.lowPrice  != null) setTodayLow( l => l == null ? data.lowPrice  : Math.min(l, data.lowPrice));
-    },
-    onPendingOrders: (data) => {
-      setWsPendingData(data);
-    },
-  });
-
-  // ── 렌더 ────────────────────────────────────────────────────
-  return (
-      <div className="h-full flex flex-col bg-stone-100 text-stone-800 overflow-hidden">
-
-        {loginModal && (
-            <LoginRequiredModal message={loginModal} onClose={() => setLoginModal(null)} />
-        )}
-
-        {/* 상단: 종목 헤더 — tokenInfo 로드 여부와 관계없이 항상 렌더 (탭 접근 보장) */}
-        <AssetHeader
-            asset={asset}
-            currentPrice={currentPrice}
-            basePrice={basePrice}
-            activeTab={activeTab}
-            onTabChange={setActiveTab}
-            isLiked={isLiked}
-            onToggleLike={async () => {
-                try {
-                    await toggleLike(TOKEN_ID);
-                } catch (err) {
-                    console.error('[MockupPage] like toggle failed:', err);
-                }
-            }}
-            hideStats
-        />
-        {/* 메인 콘텐츠 */}
-        <div className="flex-1 flex overflow-hidden p-6 gap-6">
-
-          {activeTab === 'chart' ? (
-              <>
-                {/* ────────── 차트 패널 ────────── */}
-                <div className="flex-[2] flex flex-col gap-6 overflow-hidden min-h-0">
-
-                  {/* 캔들 차트 */}
-                  <div className="flex-none bg-white rounded-2xl border border-stone-200 flex flex-col overflow-hidden shadow-sm">
-                    <div className="p-4 border-b border-stone-200 flex items-center justify-between">
-                      <div className="flex items-center gap-4">
-                        <div className="flex bg-stone-200 p-1 rounded-lg">
-                          {CHART_PERIODS.map(p => (
-                              <button
-                                  key={p}
-                                  onClick={() => setChartPeriod(p)}
-                                  className={cn(
-                                      'px-3 py-1 rounded-md text-[11px] font-bold transition-all',
-                                      chartPeriod === p
-                                          ? 'bg-white text-stone-800 shadow-sm'
-                                          : 'text-stone-400 hover:text-stone-500'
-                                  )}
-                              >
-                                {p}
-                              </button>
-                          ))}
-                        </div>
-                      </div>
-                      <button
-                          onClick={() => setChartModalOpen(true)}
-                          className="flex items-center gap-1 text-[11px] font-bold text-stone-400 hover:text-stone-800 transition-colors"
-                      >
-                        차트 크게보기 <Maximize2 size={14} />
-                      </button>
-                    </div>
-
-                    <div className="h-[260px] flex flex-col">
-                      {/* OHLCV 인디케이터 */}
-                      <div className="px-6 pt-4 pb-1 flex items-center gap-3">
-                        {[
-                          { label: '시', val: displayData?.open },
-                          { label: '고', val: displayData?.high, color: 'var(--color-brand-red)' },
-                          { label: '저', val: displayData?.low,  color: 'var(--color-brand-blue)' },
-                          { label: '종', val: displayData?.close },
-                        ].map(({ label, val, color }) => (
-                            <div key={label} className="flex items-center gap-1">
-                              <span className="text-[10px] font-bold text-stone-400">{label}</span>
-                              <span className="text-[11px] font-mono font-bold"
-                                    style={{ color: color ?? 'var(--color-stone-800)' }}>
-                                {val?.toLocaleString() ?? '-'}
-                              </span>
-                            </div>
-                        ))}
-                      </div>
-
-                      {chartData.length === 0 ? (
-                          <div className="flex-1 flex items-center justify-center text-stone-400 text-sm font-bold">
-                            {loading ? '데이터 로딩 중...' : '캔들 데이터 없음 (백엔드 확인 필요)'}
-                          </div>
-                      ) : (
-                          <>
-                            {/* 캔들 차트 */}
-                            <div className="flex-1 px-2">
-                              <ResponsiveContainer width="100%" height="100%">
-                                <ComposedChart
-                                    data={chartData}
-                                    margin={{ top: 8, right: 48, left: 0, bottom: 0 }}
-                                    onMouseMove={e => {
-                                      if (e?.activePayload?.length > 0) setHoveredData(e.activePayload[0].payload);
-                                    }}
-                                    onMouseLeave={() => setHoveredData(null)}
-                                >
-                                  <CartesianGrid strokeDasharray="3 3" stroke="#f0efee" vertical={false} />
-                                  <XAxis dataKey="time" hide />
-                                  <YAxis orientation="right"
-                                         domain={yDomain}
-                                         ticks={yTicks}
-                                         tick={{ fontSize: 10, fill: '#a8a29e', fontWeight: 'bold' }}
-                                         axisLine={false} tickLine={false}
-                                         tickFormatter={v => v.toLocaleString()} width={56} />
-                                  <Tooltip
-                                      contentStyle={{
-                                        background: '#fafaf9', border: '1px solid #e7e5e4',
-                                        borderRadius: 12, fontSize: 11, color: '#292524',
-                                        boxShadow: '0 4px 12px rgba(0,0,0,0.05)',
-                                      }}
-                                      formatter={(value, name, props) => {
-                                        const d = props?.payload;
-                                        if (!d) return [value, name];
-                                        return [
-                                          `시 ${d.open?.toLocaleString()}  고 ${d.high?.toLocaleString()}  저 ${d.low?.toLocaleString()}  종 ${d.close?.toLocaleString()}`,
-                                          d.time,
-                                        ];
-                                      }}
-                                      labelFormatter={() => ''}
-                                  />
-                                  <Bar dataKey={d => [d.low, d.high]}
-                                       shape={<CandlestickShape />} animationDuration={0} isAnimationActive={false} />
-                                </ComposedChart>
-                              </ResponsiveContainer>
-                            </div>
-
-                            {/* 거래량 차트 + x축 레이블 */}
-                            <div className="h-[72px] px-2 border-t border-stone-100">
-                              <ResponsiveContainer width="100%" height="100%">
-                                <BarChart data={chartData} margin={{ top: 2, right: 48, left: 0, bottom: 0 }}>
-                                  <XAxis dataKey="time" axisLine={false} tickLine={false}
-                                         tick={{ fontSize: 9, fill: '#a8a29e' }} minTickGap={40}
-                                         interval="preserveStartEnd" />
-                                  <YAxis hide domain={[0, 'auto']} />
-                                  <Bar dataKey="vol" radius={[2, 2, 0, 0]} isAnimationActive={false}>
-                                    {chartData.map((d, i) => (
-                                      <Cell key={i} fill={
-                                        d.close == null || d.open == null ? '#d6d3d1'
-                                        : d.close >= d.open ? 'rgba(239,68,68,0.45)'
-                                        : 'rgba(59,130,246,0.45)'
-                                      } />
-                                    ))}
-                                  </Bar>
-                                </BarChart>
-                              </ResponsiveContainer>
-                            </div>
-                          </>
-                      )}
-                    </div>
-                  </div>
-
-                  {/* 시세 섹션 */}
-                  <div className="flex-1 min-h-0 bg-white rounded-2xl border border-stone-200 flex flex-col overflow-hidden shadow-sm">
-                    <div className="p-4 border-b border-stone-200 flex items-center justify-between">
-                      <h3 className="text-sm font-bold text-stone-800">시세</h3>
-                      <span className="text-xs font-bold text-stone-400">실시간</span>
-                      <X size={14} className="text-stone-400 cursor-pointer" />
-                    </div>
-                    <div className="flex-1 overflow-y-auto">
-                      <table className="w-full text-sm">
-                        <thead className="text-stone-400 border-b border-stone-200 sticky top-0 bg-white">
-                        <tr>
-                          <th className="text-left p-4 font-bold">체결가</th>
-                          <th className="text-right p-4 font-bold">체결량(주)</th>
-                          <th className="text-right p-4 font-bold">등락률</th>
-                          <th className="text-right p-4 font-bold">거래량(주)</th>
-                          <th className="text-right p-4 font-bold">시간</th>
-                        </tr>
-                        </thead>
-                        <tbody className="divide-y divide-stone-200">
-                        {trades.map((row, i) => (
-                            <PriceRow key={i} {...row} />
-                        ))}
-                        </tbody>
-                      </table>
-                    </div>
-                  </div>
-                </div>
-
-                {/* ────────── 호가 패널 ────────── */}
-                <div className="w-[400px] min-h-0 bg-white rounded-2xl border border-stone-200 flex flex-col overflow-hidden shadow-sm text-stone-800">
-                  <div className="p-4 border-b border-stone-200 flex items-center justify-between bg-stone-100">
-                    <div className="flex items-center gap-2">
-                      <h3 className="text-sm font-black text-stone-800">호가</h3>
-                    </div>
-                    <div className="flex gap-2">
-                      <button className="p-1 hover:text-stone-800 text-stone-400">
-                        <Target size={14} />
-                      </button>
-                      <X size={14} className="text-stone-400 cursor-pointer hover:text-stone-800" />
-                    </div>
-                  </div>
-
-                  {/* 토스 스타일 3열 호가창 */}
-                  {(() => {
-                    // 체결강도 계산
-                    const buyVol  = executions.filter(e => e.isBuy).reduce((s, e) => s + e.qty, 0);
-                    const sellVol = executions.filter(e => !e.isBuy).reduce((s, e) => s + e.qty, 0);
-                    const strengthNum = sellVol > 0 ? Math.round((buyVol / sellVol) * 100) : buyVol > 0 ? 200 : null;
-                    const strengthVal = strengthNum != null ? `${strengthNum}%` : '대기중';
-                    const strengthColor = strengthNum == null ? 'text-stone-400' : strengthNum >= 100 ? 'text-brand-red' : 'text-brand-blue';
-
-                    // 매도 우측 통계 항목
-                    const statItems = [
-                      // TODO: 백엔드 로직 미구현 — 전일 종가 기준 ±30% 하드코딩
-                      { label: '상한가', value: basePrice > 0 ? Math.round(basePrice * 1.3).toLocaleString() : '-', color: 'var(--color-brand-red)', unimplemented: true },
-                      { label: '하한가', value: basePrice > 0 ? Math.round(basePrice * 0.7).toLocaleString() : '-', color: 'var(--color-brand-blue)', unimplemented: true },
-                      { label: '시작',   value: tokenInfo?.todayOpenPrice != null ? tokenInfo.todayOpenPrice.toLocaleString() : '체결 기록 없음', color: tokenInfo?.todayOpenPrice != null ? '#292524' : 'var(--color-stone-400)' },
-                      { label: '1일최고', value: todayHigh != null ? todayHigh.toLocaleString() : '체결 기록 없음', color: todayHigh != null ? 'var(--color-brand-red)' : 'var(--color-stone-400)' },
-                      { label: '1일최저', value: todayLow  != null ? todayLow.toLocaleString()  : '체결 기록 없음', color: todayLow  != null ? 'var(--color-brand-blue)' : 'var(--color-stone-400)' },
-                    ];
-
-                    // 높은 가격 → 낮은 가격 순 (현재가 바 바로 위가 가장 낮은 매도호가)
-                    // statItems 개수만큼 행이 보장되어야 모든 통계가 표시됨 — ask가 부족하면 null 패딩
-                    const reversedAsks = [...asks].reverse();
-                    const paddedAsks = reversedAsks.length >= statItems.length
-                      ? reversedAsks
-                      : [...Array(statItems.length - reversedAsks.length).fill(null), ...reversedAsks];
-
-                    return (
-                      <div className="flex-1 min-h-0 overflow-hidden flex flex-col">
-                        {/* 컬럼 헤더 */}
-                        <div className="grid grid-cols-[80px_1fr_80px] border-b border-stone-200 bg-stone-100">
-                          <span className="text-[10px] font-bold text-stone-400 text-right pr-2 py-1.5">매도수량</span>
-                          <span className="text-[10px] font-bold text-stone-400 text-center py-1.5">호가</span>
-                          <span className="text-[10px] font-bold text-stone-400 text-left pl-2 py-1.5">매수수량</span>
-                        </div>
-
-                        {/* ── 매도 영역 (상단 절반) — justify-end로 현재가 바에 붙게 ── */}
-                        <div className="flex-1 min-h-0 overflow-y-auto flex flex-col justify-end">
-                          {paddedAsks.map((row, i) => {
-                            const cp = row && basePrice > 0 ? ((row.price - basePrice) / basePrice) * 100 : 0;
-                            const dp = row && maxAskAmount > 0 ? (row.amount / maxAskAmount) * 100 : 0;
-                            const stat = statItems[i];
-                            return (
-                              <div key={`ask-${i}`} className="grid grid-cols-[80px_1fr_80px] h-9 border-b border-stone-100 hover:bg-blue-50/20 cursor-pointer">
-                                {/* 좌: 매도수량 */}
-                                <div className="relative flex items-center justify-end pr-2 overflow-hidden">
-                                  {row && <div className="absolute right-0 top-0 bottom-0 bg-brand-blue/10" style={{ width: `${dp}%` }} />}
-                                  {row && <span className="relative z-10 text-[11px] font-mono font-bold text-brand-blue">{row.amount.toLocaleString()}</span>}
-                                </div>
-                                {/* 중: 가격 + % */}
-                                <div className="flex flex-col items-center justify-center">
-                                  {row && <>
-                                    <span className="text-[12px] font-mono font-black text-brand-blue">{row.price.toLocaleString()}</span>
-                                    <span className="text-[9px] font-bold text-brand-blue/60">{cp >= 0 ? '+' : ''}{cp.toFixed(2)}%</span>
-                                  </>}
-                                </div>
-                                {/* 우: 통계 */}
-                                <div className="flex items-center pl-1 pr-2">
-                                  {stat && (
-                                    <div className="flex flex-col w-full gap-0.5">
-                                      <div className="flex justify-between w-full">
-                                        <span className="text-[9px] font-bold text-stone-400">{stat.label}</span>
-                                        <span className="text-[9px] font-bold" style={{ color: stat.color }}>{stat.value}</span>
-                                      </div>
-                                      {stat.unimplemented && (
-                                        <span className="text-[8px] font-bold text-amber-400 text-right">미구현</span>
-                                      )}
-                                    </div>
-                                  )}
-                                </div>
-                              </div>
-                            );
-                          })}
-                        </div>
-
-                        {/* ── 현재가 바 (중앙 고정) ── */}
-                        <div className="h-10 bg-stone-200 flex items-center justify-center relative flex-none border-y border-stone-300">
-                          <div className="absolute left-2 w-5 h-5 bg-brand-blue rounded flex items-center justify-center text-[9px] font-black text-white">현</div>
-                          <div className="flex items-baseline gap-2">
-                            <span className="text-sm font-black text-stone-800 font-mono tracking-tight">
-                              {currentPrice > 0 ? currentPrice.toLocaleString() : '-'}
-                            </span>
-                            {currentPrice > 0 && basePrice > 0 && (() => {
-                              const r = ((currentPrice - basePrice) / basePrice) * 100;
-                              return (
-                                <span className={`text-[10px] font-bold ${r >= 0 ? 'text-brand-red' : 'text-brand-blue'}`}>
-                                  {r >= 0 ? '+' : ''}{r.toFixed(2)}%
-                                </span>
-                              );
-                            })()}
-                          </div>
-                        </div>
-
-                        {/* ── 매수 영역 (하단 절반) ── */}
-                        <div className="flex-1 min-h-0 overflow-y-auto">
-                          {/* 체결강도 행 */}
-                          <div className="grid grid-cols-[80px_1fr_80px] h-9 border-b border-stone-100 bg-stone-50/60">
-                            <div className="flex items-center px-2">
-                              <div>
-                                <p className="text-[9px] font-bold text-stone-400 leading-none">실시간 체결강도</p>
-                                <p className={`text-[11px] font-black leading-none mt-0.5 ${strengthColor}`}>{strengthVal}</p>
-                              </div>
-                            </div>
-                            <div className="flex items-center justify-center">
-                              <span className="text-[11px] font-bold text-stone-300">—</span>
-                            </div>
-                            <div />
-                          </div>
-
-                          {/* 매수 호가 rows */}
-                          {bids.map((row, i) => {
-                            const cp = basePrice > 0 ? ((row.price - basePrice) / basePrice) * 100 : 0;
-                            const dp = maxBidAmount > 0 ? (row.amount / maxBidAmount) * 100 : 0;
-                            const ex = executions[i];
-                            return (
-                              <div key={`bid-${i}`} className="grid grid-cols-[80px_1fr_80px] h-9 border-b border-stone-100 hover:bg-red-50/20 cursor-pointer">
-                                {/* 좌: 체결 내역 */}
-                                <div className="flex items-center px-2 overflow-hidden">
-                                  {ex && (
-                                    <div className="flex justify-between w-full">
-                                      <span className="text-[10px] font-mono text-stone-400 truncate">{ex.price.toLocaleString()}</span>
-                                      <span className={`text-[10px] font-bold ml-1 flex-shrink-0 ${ex.isBuy ? 'text-brand-red' : 'text-brand-blue'}`}>{ex.qty}</span>
-                                    </div>
-                                  )}
-                                </div>
-                                {/* 중: 가격 + % */}
-                                <div className="flex flex-col items-center justify-center">
-                                  <span className="text-[12px] font-mono font-black text-brand-red">{row.price.toLocaleString()}</span>
-                                  <span className="text-[9px] font-bold text-brand-red/60">{cp >= 0 ? '+' : ''}{cp.toFixed(2)}%</span>
-                                </div>
-                                {/* 우: 매수수량 */}
-                                <div className="relative flex items-center justify-start pl-2 overflow-hidden">
-                                  <div className="absolute left-0 top-0 bottom-0 bg-brand-red/10" style={{ width: `${dp}%` }} />
-                                  <span className="relative z-10 text-[11px] font-mono font-bold text-brand-red">{row.amount.toLocaleString()}</span>
-                                </div>
-                              </div>
-                            );
-                          })}
-                        </div>
-                      </div>
-                    );
-                  })()}
-
-                  {/* 하단 요약 바 */}
-                  <div className="h-10 bg-stone-100 border-t border-stone-200 flex items-center justify-between px-4 text-[9px] font-black">
-                    <div className="flex gap-2">
-                      <span className="text-stone-400">판매대기</span>
-                      <span className="text-brand-blue">
-                    {asks.reduce((s, r) => s + r.amount, 0).toLocaleString()}
-                  </span>
-                    </div>
-                    <span className="text-stone-400">애프터마켓</span>
-                    <div className="flex gap-2">
-                  <span className="text-brand-red">
-                    {bids.reduce((s, r) => s + r.amount, 0).toLocaleString()}
-                  </span>
-                      <span className="text-stone-400">구매대기</span>
-                    </div>
-                  </div>
-                </div>
-              </>
-          ) : (
-              /* 차트·호가 외 탭 */
-              <div className="flex-1 bg-white rounded-2xl border border-stone-200 p-8 overflow-y-auto shadow-sm">
-                {activeTab === 'info'     && <InfoTab tokenInfo={tokenInfo} tokenAssetInfo={tokenAssetInfo} />}
-                {activeTab === 'dividend' && <DividendTab allocations={allocations} />}
-                {activeTab === 'news'     && <NewsTab disclosures={disclosures} />}
-              </div>
-          )}
-
-          {/* 주문창: 항상 오른쪽 고정 */}
-          <LoginGateOrderPanel
-              currentPrice={currentPrice}
-              isLoggedIn={!!user?.accessToken}
-              onLoginRequired={setLoginModal}
-              tokenId={TOKEN_ID}
-              token={user?.accessToken}
-              wsPendingData={wsPendingData}
-          />
-        </div>
-
-        {/* ── 차트 크게보기 모달 ─────────────────────────────────── */}
-        {chartModalOpen && (
-            <div
-                className="fixed inset-0 z-50 flex items-center justify-center bg-black/60 backdrop-blur-sm"
-                onClick={() => setChartModalOpen(false)}
-            >
-              <div
-                  className="bg-white rounded-2xl border border-stone-200 shadow-2xl w-[90vw] max-w-[1200px] h-[80vh] flex flex-col overflow-hidden"
-                  onClick={e => e.stopPropagation()}
-              >
-                {/* 모달 헤더 */}
-                <div className="p-4 border-b border-stone-200 flex items-center justify-between">
-                  <div className="flex items-center gap-4">
-                <span className="text-sm font-black text-stone-800">
-                  {tokenInfo?.assetName ?? '차트'}
-                </span>
-                    <div className="flex bg-stone-200 p-1 rounded-lg">
-                      {CHART_PERIODS.map(p => (
-                          <button
-                              key={p}
-                              onClick={() => setChartPeriod(p)}
-                              className={cn(
-                                  'px-3 py-1 rounded-md text-[11px] font-bold transition-all',
-                                  chartPeriod === p
-                                      ? 'bg-white text-stone-800 shadow-sm'
-                                      : 'text-stone-400 hover:text-stone-500'
-                              )}
-                          >
-                            {p}
-                          </button>
-                      ))}
-                    </div>
-                  </div>
-                  <button
-                      onClick={() => setChartModalOpen(false)}
-                      className="p-1 rounded-lg hover:bg-stone-100 transition-colors"
-                  >
-                    <X size={18} className="text-stone-500" />
-                  </button>
-                </div>
-
-                {/* 모달 차트 */}
-                <div className="flex-1 flex flex-col overflow-hidden">
-                  {/* OHLCV */}
-                  <div className="px-6 pt-4 pb-1 flex items-center gap-3">
-                    {[
-                      { label: '시', val: displayData?.open },
-                      { label: '고', val: displayData?.high, color: 'var(--color-brand-red)' },
-                      { label: '저', val: displayData?.low,  color: 'var(--color-brand-blue)' },
-                      { label: '종', val: displayData?.close },
-                    ].map(({ label, val, color }) => (
-                        <div key={label} className="flex items-center gap-1">
-                          <span className="text-[10px] font-bold text-stone-400">{label}</span>
-                          <span className="text-[11px] font-mono font-bold"
-                                style={{ color: color ?? 'var(--color-stone-800)' }}>
-                            {val?.toLocaleString() ?? '-'}
-                          </span>
-                        </div>
-                    ))}
-                  </div>
-
-                  {chartData.length === 0 ? (
-                      <div className="flex-1 flex items-center justify-center text-stone-400 text-sm font-bold">
-                        {loading ? '데이터 로딩 중...' : '캔들 데이터 없음'}
-                      </div>
-                  ) : (
-                      <>
-                        <div className="flex-1 px-2">
-                          <ResponsiveContainer width="100%" height="100%">
-                            <ComposedChart
-                                data={chartData}
-                                margin={{ top: 8, right: 60, left: 0, bottom: 0 }}
-                                onMouseMove={e => {
-                                  if (e?.activePayload?.length > 0) setHoveredData(e.activePayload[0].payload);
-                                }}
-                                onMouseLeave={() => setHoveredData(null)}
-                            >
-                              <CartesianGrid strokeDasharray="3 3" stroke="#f0efee" vertical={false} />
-                              <XAxis dataKey="time" hide />
-                              <YAxis orientation="right"
-                                     domain={([min, max]) => {
-                                       const pad = Math.max(Math.ceil((max - min) * 0.05), 1);
-                                       return [min - pad, max + pad];
-                                     }}
-                                     tickCount={5}
-                                     tick={{ fontSize: 10, fill: '#a8a29e', fontWeight: 'bold' }}
-                                     axisLine={false} tickLine={false}
-                                     tickFormatter={v => v.toLocaleString()} width={64} />
-                              <Tooltip
-                                  contentStyle={{
-                                    background: '#fafaf9', border: '1px solid #e7e5e4',
-                                    borderRadius: 12, fontSize: 11, color: '#292524',
-                                    boxShadow: '0 4px 12px rgba(0,0,0,0.05)',
-                                  }}
-                                  formatter={(value, name, props) => {
-                                    const d = props?.payload;
-                                    if (!d) return [value, name];
-                                    return [
-                                      `시 ${d.open?.toLocaleString()}  고 ${d.high?.toLocaleString()}  저 ${d.low?.toLocaleString()}  종 ${d.close?.toLocaleString()}`,
-                                      d.time,
-                                    ];
-                                  }}
-                                  labelFormatter={() => ''}
-                              />
-                              <Bar dataKey={d => [d.low, d.high]}
-                                   shape={<CandlestickShape />} animationDuration={0} isAnimationActive={false} />
-                            </ComposedChart>
-                          </ResponsiveContainer>
-                        </div>
-                        <div className="h-[72px] px-2 border-t border-stone-100">
-                          <ResponsiveContainer width="100%" height="100%">
-                            <BarChart data={chartData} margin={{ top: 2, right: 60, left: 0, bottom: 0 }}>
-                              <XAxis dataKey="time" axisLine={false} tickLine={false}
-                                     tick={{ fontSize: 9, fill: '#a8a29e' }} minTickGap={30} />
-                              <YAxis hide domain={[0, 'auto']} />
-                              <Bar dataKey="vol" radius={[2, 2, 0, 0]} isAnimationActive={false}>
-                                {chartData.map((d, i) => (
-                                  <Cell key={i} fill={
-                                    d.close == null || d.open == null ? '#d6d3d1'
-                                    : d.close >= d.open ? 'rgba(239,68,68,0.45)'
-                                    : 'rgba(59,130,246,0.45)'
-                                  } />
-                                ))}
-                              </Bar>
-                            </BarChart>
-                          </ResponsiveContainer>
-                        </div>
-                      </>
-                  )}
-                </div>
-              </div>
-            </div>
-        )}
-      </div>
-  );
-}
-
-// ── 종목정보 탭 ─────────────────────────────────────────────────
-function InfoTab({ tokenInfo, tokenAssetInfo }) {
-  if (!tokenInfo) {
-    return (
-        <div className="flex items-center justify-center h-40 text-stone-400 font-bold text-sm">
-          정보를 불러오는 중...
-        </div>
-    );
-  }
-
-  const rows = [
-    { label: '토큰 이름',    value: tokenInfo.tokenName   ?? '-' },
-    { label: '토큰 심볼',    value: tokenInfo.tokenSymbol ?? '-' },
-    { label: '총 발행량',    value: (tokenAssetInfo?.totalSupply ?? tokenInfo.totalSupply) != null ? `${(tokenAssetInfo?.totalSupply ?? tokenInfo.totalSupply).toLocaleString()}주` : '-' },
-    { label: '발행가',       value: tokenAssetInfo?.initPrice != null ? `${tokenAssetInfo.initPrice.toLocaleString()}원` : '-' },
-    { label: '총 자산 가치', value: tokenAssetInfo?.totalValue != null ? `${tokenAssetInfo.totalValue.toLocaleString()}원` : '-' },
-    { label: '자산 주소',    value: tokenAssetInfo?.assetAddress ?? '-' },
-    { label: '상장일',       value: (tokenAssetInfo?.createdAt ?? tokenInfo.issuedAt) ? new Date(tokenAssetInfo?.createdAt ?? tokenInfo.issuedAt).toLocaleDateString('ko-KR') : '-' },
-  ];
-
-  return (
-      <div className="space-y-8 max-w-4xl">
-        <section>
-          <h3 className="text-lg font-bold mb-4 text-stone-800">종목 상세 정보</h3>
-          <div className="grid grid-cols-2 gap-3">
-            {rows.map((item, i) => (
-                <div key={i} className="flex justify-between p-4 bg-stone-100 rounded-xl border border-stone-200">
-                  <span className="text-stone-400 font-bold text-sm">{item.label}</span>
-                  <span className="text-stone-800 font-bold text-sm">{item.value}</span>
-                </div>
-            ))}
-          </div>
-        </section>
-
-        {tokenInfo.assetName && (
-            <section>
-              <h3 className="text-lg font-bold mb-4 text-stone-800">자산 개요</h3>
-              <p className="text-stone-500 text-sm leading-relaxed">
-                {tokenInfo.assetName}
-                {tokenAssetInfo?.assetAddress ? ` — ${tokenAssetInfo.assetAddress}` : ''}
-              </p>
-            </section>
-        )}
-      </div>
-  );
-}
-
-// ── 배당금 탭 ───────────────────────────────────────────────────
-function DividendTab({ allocations }) {
-  if (allocations === null) {
-    return (
-        <div className="flex items-center justify-center h-40 text-stone-400 font-bold text-sm">
-          배당금 내역을 불러오는 중...
-        </div>
-    );
-  }
-
-  const totalDividend = allocations.reduce((sum, a) => sum + (a.monthlyDividendIncome ?? 0), 0);
-
-  return (
-      <div className="space-y-6">
-        <div className="flex items-center justify-between">
-          <h3 className="text-lg font-bold text-stone-800">배당금 내역</h3>
-          <div className="flex items-center gap-2 text-sm text-stone-400">
-            <span>총 누적 배당금:</span>
-            <span className="text-stone-800 font-black">
-            {totalDividend > 0 ? `${totalDividend.toLocaleString()}원` : '-'}
-          </span>
-          </div>
-        </div>
-        {allocations.length === 0 ? (
-            <div className="py-16 flex flex-col items-center gap-3 text-center text-stone-400">
-              <p className="text-sm font-bold">배당금 내역이 없습니다</p>
-            </div>
-        ) : (
-            <div className="overflow-hidden rounded-xl border border-stone-200">
-              <table className="w-full text-sm">
-                <thead className="bg-stone-100 text-stone-400 text-[11px] font-bold">
-                <tr>
-                  <th className="text-left p-4">지급일</th>
-                  <th className="text-right p-4">주당 배당금</th>
-                  <th className="text-right p-4">총 배당금</th>
-                  <th className="text-right p-4">배치 여부</th>
-                </tr>
-                </thead>
-                <tbody className="divide-y divide-stone-200">
-                {allocations.map((a, i) => (
-                    <tr key={i} className="hover:bg-stone-50">
-                      <td className="p-4 text-stone-600 font-bold">
-                        {a.settledAt ? new Date(a.settledAt).toLocaleDateString('ko-KR') : '-'}
-                      </td>
-                      <td className="p-4 text-right font-mono font-bold text-stone-800">
-                        {a.allocationPerToken != null ? `${a.allocationPerToken.toLocaleString()}원` : '-'}
-                      </td>
-                      <td className="p-4 text-right font-mono font-bold text-stone-800">
-                        {a.monthlyDividendIncome != null ? `${a.monthlyDividendIncome.toLocaleString()}원` : '-'}
-                      </td>
-                      <td className="p-4 text-right">
-                    <span className={cn(
-                        'px-2 py-0.5 rounded text-[10px] font-bold',
-                        a.allocationBatchStatus ? 'bg-green-100 text-green-700' : 'bg-stone-200 text-stone-500'
-                    )}>
-                      {a.allocationBatchStatus ? '완료' : '대기'}
-                    </span>
-                      </td>
-                    </tr>
-                ))}
-                </tbody>
-              </table>
-            </div>
-        )}
-      </div>
-  );
-}
-
-// ── 공시 탭 ─────────────────────────────────────────────────────
-const DISCLOSURE_CATEGORY_LABEL = {
-  BUILDING:  '건물',
-  DIVIDEND:  '배당',
-  ETC:       '기타',
-};
-
-function NewsTab({ disclosures }) {
-  const [expanded, setExpanded] = useState(null);
-
-  if (disclosures === null) {
-    return (
-        <div className="flex items-center justify-center h-40 text-stone-400 font-bold text-sm">
-          공시 내역을 불러오는 중...
-        </div>
-    );
-  }
-
-  return (
-      <div className="space-y-4">
-        <h3 className="text-lg font-bold text-stone-800">공시</h3>
-        {disclosures.length === 0 ? (
-            <div className="py-16 flex flex-col items-center gap-3 text-center text-stone-400">
-              <p className="text-sm font-bold">공시 내역이 없습니다</p>
-            </div>
-        ) : (
-            <div className="space-y-2">
-              {disclosures.map((d, i) => (
-                  <div key={i} className="border border-stone-200 rounded-xl overflow-hidden">
-                    <button
-                        className="w-full flex items-center justify-between p-4 hover:bg-stone-50 transition-colors text-left"
-                        onClick={() => setExpanded(expanded === i ? null : i)}
-                    >
-                      <div className="flex items-center gap-3">
-                  <span className="px-2 py-0.5 bg-stone-200 rounded text-[10px] font-bold text-stone-500">
-                    {DISCLOSURE_CATEGORY_LABEL[d.disclosureCategory] ?? d.disclosureCategory}
-                  </span>
-                        <span className="text-sm font-bold text-stone-800">{d.disclosureTitle}</span>
-                      </div>
-                      <div className="flex items-center gap-3 shrink-0">
-                  <span className="text-[11px] text-stone-400 font-bold">
-                    {d.createdAt ? new Date(d.createdAt).toLocaleDateString('ko-KR') : '-'}
-                  </span>
-                        <span className="text-stone-400 text-xs">{expanded === i ? '▲' : '▼'}</span>
-                      </div>
-                    </button>
-                    {expanded === i && (
-                        <div className="px-4 pb-4 pt-2 border-t border-stone-100 bg-stone-50">
-                          <p className="text-sm text-stone-600 leading-relaxed whitespace-pre-wrap">
-                            {d.disclosureContent}
-                          </p>
-                          {d.OriginName && (
-                              <p className="mt-2 text-[11px] text-stone-400 font-bold">첨부: {d.OriginName}</p>
-                          )}
-                        </div>
-                    )}
-                  </div>
-              ))}
-            </div>
-        )}
-      </div>
-  );
-}
-
-// ── 로그인 게이트 주문창 ────────────────────────────────────────
-// 비로그인 시 매수/매도/대기 버튼에 로그인 안내 처리
-function LoginGateOrderPanel({ currentPrice, isLoggedIn, onLoginRequired, tokenId, token, wsPendingData }) {
-  const [orderSide, setOrderSide] = useState('buy');
-  const [inputMode, setInputMode] = useState('qty');
-  const [price, setPrice]             = useState(currentPrice > 0 ? String(currentPrice) : '');
-  const [qty, setQty]                 = useState('');
-  const [accountPassword, setAccountPassword] = useState('');
-
-  // currentPrice 로드 완료 시 가격 필드 자동 세팅 (비어있을 때만)
-  useEffect(() => {
-    if (currentPrice > 0) setPrice(p => p === '' ? String(currentPrice) : p);
-  }, [currentPrice]);
-  const [amount, setAmount]           = useState('');
-  const [submitting, setSubmitting]   = useState(false);
-  const [orderMsg, setOrderMsg]       = useState(null); // { type: 'success'|'error', text }
-
-  // ── 대기 주문 목록 ───────────────────────────────────────────
-  const [pendingOrders, setPendingOrders] = useState([]);
-  const [pendingLoading, setPendingLoading] = useState(false);
-
-  // ── 주문 수정 상태 ───────────────────────────────────────────
-  const [editingOrderId, setEditingOrderId] = useState(null);
-  const [editPrice, setEditPrice]           = useState('');
-  const [editQty, setEditQty]               = useState('');
-  const [editAccountPassword, setEditAccountPassword] = useState('');
-  const [updateMsg, setUpdateMsg]           = useState(null); // { orderId, type, text }
-  const [capacity, setCapacity]             = useState({ availableBalance: 0, availableQuantity: 0 });
-  const [passwordModal, setPasswordModal]   = useState(null); // { action, orderId? }
-  const [confirmModal, setConfirmModal]     = useState(null); // { action, orderId? }
-
-  // WS 실시간 업데이트 수신 시 목록 교체 (편집 중인 주문은 유지)
-  useEffect(() => {
-    if (!wsPendingData) return;
-
-    if (editingOrderId === null) {
-      setPendingOrders(wsPendingData);
-      return;
-    }
-
-    const editingOrderExists = wsPendingData.some(o => o.orderId === editingOrderId);
-
-    if (!editingOrderExists) {
-      setPendingOrders(wsPendingData);
-      setEditingOrderId(null);
-      setEditPrice('');
-      setEditQty('');
-      setUpdateMsg({
-        orderId: editingOrderId,
-        type: 'error',
-        text: '편집 중인 주문이 체결되었거나 취소되어 편집이 종료되었습니다.',
-      });
-      return;
-    }
-
-    setPendingOrders(prev =>
-        wsPendingData.map(incoming => {
-          if (incoming.orderId === editingOrderId) {
-            return prev.find(o => o.orderId === editingOrderId) ?? incoming;
-          }
-          return incoming;
-        })
-    );
-  }, [wsPendingData, editingOrderId]);
-
-  const fetchPendingOrders = useCallback(async () => {
-    if (!isLoggedIn || !token) return;
-    setPendingLoading(true);
-    try {
-      const res = await api.get(`/api/token/${tokenId}/order/pending`);
-      setPendingOrders(res.data);
-    } catch (e) {
-      console.warn('[OrderPanel] 대기 주문 조회 실패:', e.message);
-    } finally {
-      setPendingLoading(false);
-    }
-  }, [isLoggedIn, token, tokenId]);
-
-  useEffect(() => {
-    if (orderSide === 'pending') fetchPendingOrders();
-  }, [orderSide, fetchPendingOrders]);
-
-  // 가용 잔고/수량 조회
-  const fetchCapacity = useCallback(async () => {
-    if (!isLoggedIn || !tokenId) return;
-    try {
-      const res = await api.get(`/api/token/${tokenId}/order/capacity`);
-      setCapacity({
-        availableBalance:  res.data.availableBalance  ?? 0,
-        availableQuantity: res.data.availableQuantity ?? 0,
-      });
-    } catch (e) {
-      console.warn('[OrderPanel] capacity 조회 실패:', e.message);
-    }
-  }, [isLoggedIn, token, tokenId]);
-
-  useEffect(() => {
-    if (orderSide === 'buy' || orderSide === 'sell') fetchCapacity();
-  }, [orderSide, fetchCapacity]);
-
-  const isBuy     = orderSide === 'buy';
-  const isPending = orderSide === 'pending';
-
-  const numPrice  = Number(price) || 0;
-  const numQty    = inputMode === 'qty'
-      ? (Number(qty) || 0)
-      : (numPrice > 0 ? Math.floor((Number(amount) || 0) / numPrice) : 0);
-  const numAmount = inputMode === 'amount'
-      ? (Number(amount) || 0)
-      : numPrice * numQty;
-
-  function isValidAccountPassword(value) {
-    return /^\d{4}$/.test(value);
-  }
-
-  function getApiErrorMessage(error, fallback) {
-    return error?.response?.data?.errorMessage || error?.response?.data?.message || error?.message || fallback;
-  }
-
-  function openPasswordModal(action, orderId = null) {
-    setPasswordModal({ action, orderId });
-    setConfirmModal(null);
-    if (action === 'create') {
-      setAccountPassword('');
-      setOrderMsg(null);
-    } else {
-      setEditAccountPassword('');
-      setUpdateMsg(null);
-    }
-  }
-
-  function closePasswordModal() {
-    setPasswordModal(null);
-    setAccountPassword('');
-    setEditAccountPassword('');
-  }
-
-  function resetOrderAuthFlow() {
-    setPasswordModal(null);
-    setConfirmModal(null);
-    setAccountPassword('');
-    setEditAccountPassword('');
-  }
-
-  async function verifyPasswordBeforeConfirm() {
-    const password = passwordModal?.action === 'create' ? accountPassword : editAccountPassword;
-
-    try {
-      await api.post('/api/myaccount/verify-password', { accountPassword: password });
-      return { ok: true };
-    } catch (e) {
-      return { ok: false, message: getApiErrorMessage(e, '계좌 비밀번호를 다시 확인해 주세요.') };
-    }
-  }
-
-  function handleTabClick(side) {
-    if (!isLoggedIn && side === 'pending') {
-      onLoginRequired('로그인해야 볼 수 있습니다');
-      return;
-    }
-    setOrderMsg(null);
-    setOrderSide(side);
-  }
-
-  async function handleSubmit() {
-    if (!isLoggedIn) {
-      onLoginRequired('매수/매도 주문을 하려면\n먼저 로그인해야 합니다');
-      return;
-    }
-    if (!Number.isInteger(numPrice) || !Number.isInteger(numQty) || numPrice <= 0 || numQty <= 0) {
-      setOrderMsg({ type: 'error', text: '가격과 수량은 양의 정수만 입력하세요.' });
-      return;
-    }
-    const tick = getTickSize(numPrice);
-    if (numPrice % tick !== 0) {
-      setOrderMsg({ type: 'error', text: `호가 단위에 맞지 않습니다. ${numPrice.toLocaleString()}원 근처 호가 단위는 ${tick.toLocaleString()}원입니다.` });
-      return;
-    }
-    if (false && !isValidAccountPassword(accountPassword)) {
-      setOrderMsg({ type: 'error', text: '계좌 비밀번호 4자리를 입력해 주세요.' });
-      return;
-    }
-    openPasswordModal('create');
-    return;
-    setSubmitting(true);
-    setOrderMsg(null);
-    try {
-      const body = {
-        orderPrice:    numPrice,
-        orderQuantity: numQty,
-        orderType:     isBuy ? 'BUY' : 'SELL',
-        accountPassword,
-      };
-      await api.post(`/api/token/${tokenId}/order`, body);
-      setOrderMsg({ type: 'success', text: `${isBuy ? '매수' : '매도'} 주문이 접수되었습니다.` });
-      setQty('');
-      setAmount('');
-      setAccountPassword('');
-      fetchCapacity();
-    } catch (e) {
-      setOrderMsg({ type: 'error', text: getApiErrorMessage(e, '주문 접수에 실패했습니다.') });
-    } finally {
-      setSubmitting(false);
-    }
-  }
-
-  async function submitCreateOrder() {
-    if (!isValidAccountPassword(accountPassword)) {
-      setOrderMsg({ type: 'error', text: '계좌 비밀번호 4자리를 입력해 주세요.' });
-      return;
-    }
-    setSubmitting(true);
-    setOrderMsg(null);
-    try {
-      const body = {
-        orderPrice:    numPrice,
-        orderQuantity: numQty,
-        orderType:     isBuy ? 'BUY' : 'SELL',
-        accountPassword,
-      };
-      await api.post(`/api/token/${tokenId}/order`, body);
-      setOrderMsg({ type: 'success', text: `${isBuy ? '매수' : '매도'} 주문이 접수되었습니다.` });
-      setQty('');
-      setAmount('');
-      resetOrderAuthFlow();
-      fetchCapacity();
-    } catch (e) {
-      setOrderMsg({ type: 'error', text: getApiErrorMessage(e, '주문 접수에 실패했습니다.') });
-    } finally {
-      setSubmitting(false);
-    }
-  }
-
-  function handleCancelOrder(orderId) {
-    openPasswordModal('cancel', orderId);
-  }
-
-  async function submitCancelOrder(orderId) {
-    if (!token) return;
-    if (false && !isValidAccountPassword(editAccountPassword)) {
-      setUpdateMsg({ orderId, type: 'error', text: '계좌 비밀번호 4자리를 입력해 주세요.' });
-      return;
-    }
-    try {
-      await api.delete(`/api/token/order/cancel/${orderId}`, {
-        data: { accountPassword: editAccountPassword },
-      });
-      setPendingOrders(prev => prev.filter(o => o.orderId !== orderId));
-      resetOrderAuthFlow();
-    } catch (e) {
-      console.warn('[OrderPanel] 주문 취소 실패:', e.message);
-    }
-  }
-
-  function handleEditStart(o) {
-    setEditingOrderId(o.orderId);
-    setEditPrice(String(o.orderPrice ?? ''));
-    setEditQty(String(o.orderQuantity ?? ''));
-    setEditAccountPassword('');
-    setUpdateMsg(null);
-  }
-
-  function handleEditCancel() {
-    setEditingOrderId(null);
-    setEditPrice('');
-    setEditQty('');
-    setEditAccountPassword('');
-    setUpdateMsg(null);
-  }
-
-  async function handleUpdateOrder(orderId) {
-    const p = Number(editPrice);
-    const q = Number(editQty);
-    if (false && !isValidAccountPassword(editAccountPassword)) {
-      setUpdateMsg({ orderId, type: 'error', text: '계좌 비밀번호 4자리를 입력해 주세요.' });
-      return;
-    }
-    if (!Number.isInteger(p) || !Number.isInteger(q) || p <= 0 || q <= 0) {
-      setUpdateMsg({ orderId, type: 'error', text: '가격과 수량은 양의 정수만 입력하세요.' });
-      return;
-    }
-    const tick = getTickSize(p);
-    if (p % tick !== 0) {
-      setUpdateMsg({ orderId, type: 'error', text: `호가 단위에 맞지 않습니다. 이 가격대 호가 단위: ${tick.toLocaleString()}원` });
-      return;
-    }
-    try {
-      await api.put(`/api/token/order/update/${orderId}`, {
-        updatePrice: p,
-        updateQuantity: q,
-        accountPassword: editAccountPassword,
-      });
-      setPendingOrders(prev =>
-          prev.map(o => {
-            if (o.orderId !== orderId) return o;
-            const filledQuantity = Number(o.filledQuantity) || 0;
-            const remainingQuantity = Math.max(q - filledQuantity, 0);
-            return { ...o, orderPrice: p, remainingQuantity, orderQuantity: q };
-          })
-      );
-      setEditingOrderId(null);
-      setEditPrice('');
-      setEditQty('');
-      setEditAccountPassword('');
-      setUpdateMsg(null);
-    } catch (e) {
-      setUpdateMsg({ orderId, type: 'error', text: getApiErrorMessage(e, '수정에 실패했습니다.') });
-    }
-  }
-
-  // ── TickSizePolicy (백엔드와 동일한 로직) ────────────────────
-  async function submitUpdateOrder(orderId) {
-    const p = Number(editPrice);
-    const q = Number(editQty);
-    if (!Number.isInteger(p) || !Number.isInteger(q) || p <= 0 || q <= 0) {
-      setUpdateMsg({ orderId, type: 'error', text: '가격과 수량을 올바르게 입력해 주세요.' });
-      return;
-    }
-    const tick = getTickSize(p);
-    if (p % tick !== 0) {
-      setUpdateMsg({ orderId, type: 'error', text: `호가 단위를 확인해 주세요. 현재 가격의 호가 단위는 ${tick.toLocaleString()}원입니다.` });
-      return;
-    }
-    if (!isValidAccountPassword(editAccountPassword)) {
-      setUpdateMsg({ orderId, type: 'error', text: '계좌 비밀번호 4자리를 입력해 주세요.' });
-      return;
-    }
-    try {
-      await api.put(`/api/token/order/update/${orderId}`, {
-        updatePrice: p,
-        updateQuantity: q,
-        accountPassword: editAccountPassword,
-      });
-      setPendingOrders(prev =>
-          prev.map(o => {
-            if (o.orderId !== orderId) return o;
-            const filledQuantity = Number(o.filledQuantity) || 0;
-            const remainingQuantity = Math.max(q - filledQuantity, 0);
-            return { ...o, orderPrice: p, remainingQuantity, orderQuantity: q };
-          })
-      );
-      setEditingOrderId(null);
-      setEditPrice('');
-      setEditQty('');
-      resetOrderAuthFlow();
-      setUpdateMsg(null);
-    } catch (e) {
-      setUpdateMsg({ orderId, type: 'error', text: getApiErrorMessage(e, '주문 수정에 실패했습니다.') });
-    }
-  }
-
-  async function handlePasswordConfirm() {
-    if (!passwordModal) return;
-    const result = await verifyPasswordBeforeConfirm();
-    if (!result.ok) {
-      if (passwordModal.action === 'create') {
-        setAccountPassword('');
-        setOrderMsg({ type: 'error', text: result.message });
-      } else {
-        setEditAccountPassword('');
-        setUpdateMsg({ orderId: passwordModal.orderId, type: 'error', text: result.message });
-      }
-      return;
-    }
-
-    setPasswordModal(null);
-    setConfirmModal(passwordModal);
-  }
-
-  function getConfirmModalSpec() {
-    if (!confirmModal) return null;
-
-    if (confirmModal.action === 'create') {
-      return {
-        title: `${isBuy ? '매수' : '매도'} 주문 최종 확인`,
-        items: [
-          { label: '구분', value: isBuy ? '매수' : '매도' },
-          { label: '주문가', value: `${numPrice.toLocaleString()}원` },
-          { label: '주문수량', value: `${numQty.toLocaleString()}주` },
-          { label: '총 주문금액', value: `${numAmount.toLocaleString()}원` },
-        ],
-      };
-    }
-
-    const order = pendingOrders.find(o => o.orderId === confirmModal.orderId);
-    const updateTotalAmount = (Number(editPrice) || 0) * (Number(editQty) || 0);
-
-    if (confirmModal.action === 'update') {
-      return {
-        title: '주문 수정 최종 확인',
-        items: [
-          { label: '주문번호', value: String(confirmModal.orderId ?? '-') },
-          { label: '구분', value: order?.orderType === 'BUY' ? '매수' : '매도' },
-          { label: '수정가', value: `${(Number(editPrice) || 0).toLocaleString()}원` },
-          { label: '수정수량', value: `${(Number(editQty) || 0).toLocaleString()}주` },
-          { label: '주문금액', value: `${updateTotalAmount.toLocaleString()}원` },
-        ],
-      };
-    }
-
-    return {
-      title: '주문 취소 최종 확인',
-      items: [
-        { label: '주문번호', value: String(confirmModal.orderId ?? '-') },
-        { label: '구분', value: order?.orderType === 'BUY' ? '매수' : '매도' },
-        { label: '주문가', value: `${Number(order?.orderPrice || 0).toLocaleString()}원` },
-        { label: '잔여수량', value: `${Number(order?.remainingQuantity || 0).toLocaleString()}주` },
-      ],
-    };
-  }
-
-  async function handleConfirmExecution() {
-    if (!confirmModal) return;
-    if (confirmModal.action === 'create') {
-      await submitCreateOrder();
-      return;
-    }
-    if (confirmModal.action === 'update') {
-      await submitUpdateOrder(confirmModal.orderId);
-      return;
-    }
-    await submitCancelOrder(confirmModal.orderId);
-  }
-
-  function getTickSize(p) {
-    if (p < 100)   return 10;
-    if (p < 1000)  return 50;
-    if (p < 10000) return 100;
-    return 500;
-  }
-
-  const RATIO_MAP = { '10%': 0.1, '25%': 0.25, '50%': 0.5, '최대': 1.0 };
-
-  function handleRatioClick(label) {
-    const pct = RATIO_MAP[label];
-    const newQty = isBuy
-        ? (numPrice > 0 ? Math.floor(capacity.availableBalance * pct / numPrice) : 0)
-        : Math.floor(capacity.availableQuantity * pct);
-    setQty(String(newQty));
-    setInputMode('qty');
-  }
-
-  return (
-      <div className="w-[360px] min-h-0 bg-white rounded-lg border border-stone-200 flex flex-col overflow-hidden">
-
-        {/* 탭: 매수 / 매도 / 대기 */}
-        <div className="flex border-b border-stone-200">
-          {[
-            { id: 'buy',     label: '매수',  active: 'text-brand-red border-b-2 border-brand-red bg-brand-red-light/40' },
-            { id: 'sell',    label: '매도',  active: 'text-brand-blue border-b-2 border-brand-blue bg-brand-blue-light/60' },
-            { id: 'pending', label: '대기',  active: 'text-stone-800 border-b-2 border-stone-800 bg-stone-100/60' },
-          ].map(t => (
-              <button
-                  key={t.id}
-                  onClick={() => handleTabClick(t.id)}
-                  className={cn(
-                      'flex-1 py-4 text-sm font-black transition-all',
-                      orderSide === t.id ? t.active : 'text-stone-400 hover:text-stone-500'
-                  )}
-              >
-                {t.label}
-              </button>
-          ))}
-        </div>
-
-        <div className="flex-1 overflow-y-auto p-5">
-
-          {/* 대기 탭 */}
-          {isPending ? (
-              <div className="space-y-3">
-                <div className="flex items-center justify-between">
-                  <span className="text-xs font-bold text-stone-400">미체결 주문</span>
-                  <button
-                      onClick={fetchPendingOrders}
-                      disabled={pendingLoading}
-                      className="text-[10px] font-bold text-stone-400 hover:text-stone-700 transition-colors"
-                  >
-                    {pendingLoading ? '조회 중...' : '새로고침'}
-                  </button>
-                </div>
-                {pendingLoading && pendingOrders.length === 0 ? (
-                    <div className="py-16 flex flex-col items-center gap-3 text-center">
-                      <div className="w-12 h-12 bg-stone-200 rounded-lg flex items-center justify-center">
-                        <MoreHorizontal size={24} className="text-stone-400 animate-pulse" />
-                      </div>
-                      <p className="text-sm font-bold text-stone-400">조회 중...</p>
-                    </div>
-                ) : pendingOrders.length === 0 ? (
-                    <div className="py-16 flex flex-col items-center gap-3 text-center">
-                      <div className="w-12 h-12 bg-stone-200 rounded-lg flex items-center justify-center">
-                        <MoreHorizontal size={24} className="text-stone-400" />
-                      </div>
-                      <p className="text-sm font-bold text-stone-400">대기 중인 주문이 없습니다</p>
-                    </div>
-                ) : (
-                    <div className="space-y-3">
-                      {pendingOrders.map(o => {
-                        const isBuy      = o.orderType === 'BUY';
-                        const isPending  = o.orderStatus === 'PENDING';
-                        const isEditing  = editingOrderId === o.orderId;
-                        const totalAmount = isEditing
-                            ? (Number(editPrice) || 0) * (Number(editQty) || 0)
-                            : (o.orderPrice ?? 0) * (o.orderQuantity ?? 0);
-                        return (
-                            <div key={o.orderId} className="p-4 bg-stone-100 rounded-lg border border-stone-200 space-y-3">
-                              {/* 헤더: 매수/매도 뱃지 + 시간 */}
-                              <div className="flex items-center justify-between">
-                                <div className="flex items-center gap-2">
-                          <span className={cn(
-                              'text-[10px] font-black px-2 py-0.5 rounded-md',
-                              isBuy ? 'bg-brand-red-light text-brand-red' : 'bg-brand-blue-light text-brand-blue'
-                          )}>
-                            {isBuy ? '매수' : '매도'}
-                          </span>
-                                  <span className={cn(
-                                      'text-[10px] font-black px-2 py-0.5 rounded-md',
-                                      isPending
-                                          ? 'bg-stone-200 text-stone-400'
-                                          : isEditing
-                                              ? 'bg-blue-100 text-blue-600'
-                                              : 'bg-[#fef6dc] text-[#a07828]'
-                                  )}>
-                            {isPending ? '처리중' : isEditing ? '수정중' : '대기'}
-                          </span>
-                                </div>
-                                <span className="text-[9px] text-stone-400 font-bold">
-                          {o.createdAt ? new Date(o.createdAt).toLocaleTimeString('ko-KR', { hour: '2-digit', minute: '2-digit' }) : '-'}
-                        </span>
-                              </div>
-
-                              {/* 수정 모드 */}
-                              {isEditing ? (
-                                  <div className="space-y-2">
-                                    <div className="space-y-1">
-                                      <div className="flex items-center justify-between">
-                                        <label className="text-[10px] font-bold text-stone-400">수정 가격</label>
-                                        {Number(editPrice) > 0 && (
-                                          <span className="text-[10px] font-bold text-stone-400">
-                                            호가 단위: {getTickSize(Number(editPrice)).toLocaleString()}원
-                                          </span>
-                                        )}
-                                      </div>
-                                      <div className="flex items-center gap-2 bg-white border border-stone-300 rounded-md px-3 py-2">
-                                        <input
-                                            type="number"
-                                            min="1"
-                                            step={Number(editPrice) > 0 ? getTickSize(Number(editPrice)) : 1}
-                                            value={editPrice}
-                                            onChange={e => setEditPrice(e.target.value)}
-                                            className="flex-1 bg-transparent text-[11px] font-mono font-bold outline-none text-right text-stone-800"
-                                        />
-                                        <span className="text-[11px] font-bold text-stone-400">원</span>
-                                      </div>
-                                      {Number(editPrice) > 0 && Number(editPrice) % getTickSize(Number(editPrice)) !== 0 && (
-                                        <p className="text-[10px] font-bold text-amber-600">
-                                          호가 단위({getTickSize(Number(editPrice)).toLocaleString()}원)에 맞지 않습니다
-                                        </p>
-                                      )}
-                                    </div>
-                                    <div className="space-y-1">
-                                      <label className="text-[10px] font-bold text-stone-400">수정 수량</label>
-                                      <div className="flex items-center gap-2 bg-white border border-stone-300 rounded-md px-3 py-2">
-                                        <input
-                                            type="number"
-                                            min="1"
-                                            step="1"
-                                            value={editQty}
-                                            onChange={e => setEditQty(e.target.value)}
-                                            className="flex-1 bg-transparent text-[11px] font-mono font-bold outline-none text-right text-stone-800"
-                                        />
-                                        <span className="text-[11px] font-bold text-stone-400">주</span>
-                                      </div>
-                                    </div>
-                                    <div className="hidden">
-                                      <label className="text-[10px] font-bold text-stone-400">계좌 비밀번호</label>
-                                      <div className="flex items-center gap-2 bg-white border border-stone-300 rounded-md px-3 py-2">
-                                        <input
-                                            type="password"
-                                            inputMode="numeric"
-                                            maxLength={4}
-                                            value={editAccountPassword}
-                                            onChange={e => setEditAccountPassword(e.target.value.replace(/\D/g, '').slice(0, 4))}
-                                            placeholder="숫자 4자리"
-                                            className="flex-1 bg-transparent text-[11px] font-mono font-bold outline-none text-right text-stone-800 placeholder-stone-400"
-                                        />
-                                      </div>
-                                    </div>
-                                    <div className="flex justify-between text-[11px] font-bold border-t border-stone-200 pt-1.5">
-                                      <span className="text-stone-400">주문금액</span>
-                                      <span className="font-mono font-black text-stone-800">{totalAmount.toLocaleString()}원</span>
-                                    </div>
-                                    {updateMsg?.orderId === o.orderId && (
-                                        <p className={cn(
-                                            'text-[10px] font-bold text-center',
-                                            updateMsg.type === 'error' ? 'text-brand-red' : 'text-green-600'
-                                        )}>
-                                          {updateMsg.text}
-                                        </p>
-                                    )}
-                                    <div className="flex gap-2 pt-1">
-                                      <button
-                                          onClick={() => openPasswordModal('update', o.orderId)}
-                                          className="flex-1 py-2 bg-stone-800 border border-stone-800 rounded-md text-[11px] font-black text-white hover:bg-stone-700 transition-all"
-                                      >
-                                        확인
-                                      </button>
-                                      <button
-                                          onClick={handleEditCancel}
-                                          className="flex-1 py-2 bg-white border border-stone-200 rounded-md text-[11px] font-black text-stone-500 hover:bg-stone-200 transition-all"
-                                      >
-                                        취소
-                                      </button>
-                                    </div>
-                                  </div>
-                              ) : (
-                                  <>
-                                    {/* 일반 표시 모드 */}
-                                    <div className="space-y-1.5 text-[11px] font-bold">
-                                      <div className="flex justify-between">
-                                        <span className="text-stone-400">지정가격</span>
-                                        <span className="font-mono text-stone-800">{o.orderPrice?.toLocaleString()}원</span>
-                                      </div>
-                                      <div className="flex justify-between">
-                                        <span className="text-stone-400">미체결/전체</span>
-                                        <span className="font-mono text-stone-800">{o.remainingQuantity} / {o.orderQuantity}주</span>
-                                      </div>
-                                      <div className="flex justify-between border-t border-stone-200 pt-1.5">
-                                        <span className="text-stone-400">주문금액</span>
-                                        <span className="font-mono font-black text-stone-800">{totalAmount.toLocaleString()}원</span>
-                                      </div>
-                                    </div>
-                                    <div className="flex gap-2 pt-1">
-                                      <button
-                                          disabled={isPending}
-                                          onClick={() => !isPending && handleEditStart(o)}
-                                          className={cn(
-                                              'flex-1 py-2 border rounded-md text-[11px] font-black transition-all flex items-center justify-center gap-1',
-                                              isPending
-                                                  ? 'bg-stone-100 border-stone-200 text-stone-300 cursor-not-allowed'
-                                                  : 'bg-white border-stone-200 text-stone-500 hover:bg-stone-200'
-                                          )}
-                                      >
-                                        <Edit3 size={12} /> 수정
-                                      </button>
-                                      <button
-                                          disabled={isPending}
-                                          onClick={() => !isPending && handleCancelOrder(o.orderId)}
-                                          className={cn(
-                                              'flex-1 py-2 border rounded-md text-[11px] font-black transition-all flex items-center justify-center gap-1',
-                                              isPending
-                                                  ? 'bg-stone-100 border-stone-200 text-stone-300 cursor-not-allowed'
-                                                  : 'bg-brand-red-light border-brand-red-light text-brand-red hover:bg-[#fccfcf]'
-                                          )}
-                                      >
-                                        <X size={12} /> 취소
-                                      </button>
-                                    </div>
-                                  </>
-                              )}
-                            </div>
-                        );
-                      })}
-                    </div>
-                )}
-              </div>
-          ) : (
-              <div className="space-y-5">
-                {/* 주문 유형(지정가 고정) + 입력 모드 */}
-                <div className="flex items-center justify-between">
-              <span className="px-3 py-1 bg-stone-200 rounded-lg text-[10px] font-bold text-stone-500">
-                지정가
-              </span>
-                  <div className="flex bg-stone-200 p-1 rounded-lg">
-                    {[{ id: 'qty', label: '수량' }, { id: 'amount', label: '금액' }].map(m => (
-                        <button key={m.id} onClick={() => setInputMode(m.id)}
-                                className={cn(
-                                    'px-3 py-1 rounded-md text-[10px] font-bold transition-all',
-                                    inputMode === m.id ? 'bg-white text-stone-800 shadow-sm' : 'text-stone-400'
-                                )}>
-                          {m.label}
-                        </button>
-                    ))}
-                  </div>
-                </div>
-
-                {/* 가격 입력 */}
-                <div className="space-y-1.5">
-                  <div className="flex items-center justify-between">
-                    <label className="text-[10px] font-bold text-stone-400">
-                      {isBuy ? '매수' : '매도'} 가격
-                    </label>
-                    {numPrice > 0 && (
-                      <span className="text-[10px] font-bold text-stone-400">
-                        호가 단위: {getTickSize(numPrice).toLocaleString()}원
-                      </span>
-                    )}
-                  </div>
-                  <div className="flex items-center gap-2 bg-stone-200 border border-stone-200 rounded-md px-4 py-2.5">
-                    <input
-                        type="text"
-                        value={price}
-                        onChange={e => setPrice(e.target.value)}
-                        placeholder="가격 입력"
-                        className="flex-1 bg-transparent text-sm font-mono font-bold outline-none text-right pr-2 text-stone-800 placeholder-stone-400"
-                    />
-                    <span className="text-sm font-bold text-stone-400">원</span>
-                  </div>
-                  {numPrice > 0 && numPrice % getTickSize(numPrice) !== 0 && (
-                    <p className="text-[10px] font-bold text-amber-600">
-                      호가 단위({getTickSize(numPrice).toLocaleString()}원)에 맞지 않습니다
-                    </p>
-                  )}
-                </div>
-
-                {/* 수량 / 금액 입력 */}
-                {inputMode === 'qty' ? (
-                    <div className="space-y-1.5">
-                      <label className="text-[10px] font-bold text-stone-400">
-                        {isBuy ? '매수' : '매도'} 수량
-                      </label>
-                      <div className="flex items-center gap-2 bg-stone-200 border border-stone-200 rounded-md px-4 py-2.5">
-                        <input type="text" placeholder="수량 입력" value={qty}
-                               onChange={e => setQty(e.target.value)}
-                               className="flex-1 bg-transparent text-sm font-mono font-bold outline-none text-right pr-2 text-stone-800" />
-                        <span className="text-sm font-bold text-stone-400">주</span>
-                      </div>
-                    </div>
-                ) : (
-                    <div className="space-y-1.5">
-                      <label className="text-[10px] font-bold text-stone-400">
-                        {isBuy ? '매수' : '매도'} 금액
-                      </label>
-                      <div className="flex items-center gap-2 bg-stone-200 border border-stone-200 rounded-md px-4 py-2.5">
-                        <input type="text" placeholder="금액 입력" value={amount}
-                               onChange={e => setAmount(e.target.value)}
-                               className="flex-1 bg-transparent text-sm font-mono font-bold outline-none text-right pr-2 text-stone-800" />
-                        <span className="text-sm font-bold text-stone-400">원</span>
-                      </div>
-                      {amount && numPrice > 0 && (
-                          <p className="text-[10px] text-stone-400 font-bold text-right">
-                            ≈ {Math.floor(Number(amount) / numPrice)}주
-                          </p>
-                      )}
-                    </div>
-                )}
-
-                {/* 비율 버튼 */}
-                <div className="hidden">
-                  <label className="text-[10px] font-bold text-stone-400">계좌 비밀번호</label>
-                  <div className="flex items-center gap-2 bg-stone-200 border border-stone-200 rounded-md px-4 py-2.5">
-                    <input
-                        type="password"
-                        inputMode="numeric"
-                        maxLength={4}
-                        placeholder="숫자 4자리"
-                        value={accountPassword}
-                        onChange={e => setAccountPassword(e.target.value.replace(/\D/g, '').slice(0, 4))}
-                        className="flex-1 bg-transparent text-sm font-mono font-bold outline-none text-right pr-2 text-stone-800 placeholder-stone-400"
-                    />
-                  </div>
-                </div>
-
-                <div className="grid grid-cols-4 gap-2">
-                  {['10%', '25%', '50%', '최대'].map(p => (
-                      <button key={p}
-                              onClick={() => handleRatioClick(p)}
-                              disabled={!isLoggedIn}
-                              className={cn(
-                                  'py-1.5 rounded-lg text-[10px] font-bold transition-all border',
-                                  isLoggedIn
-                                      ? 'bg-stone-200 hover:bg-stone-300 border-stone-200 text-stone-400'
-                                      : 'bg-stone-100 border-stone-100 text-stone-300 cursor-not-allowed'
-                              )}>
-                        {p}
-                      </button>
-                  ))}
-                </div>
-
-                {/* 주문 요약 */}
-                <div className="pt-3 border-t border-stone-200 space-y-2">
-                  <div className="flex justify-between text-[10px] font-bold">
-                    <span className="text-stone-400">{isBuy ? '매수가능 금액' : '매도가능 수량'}</span>
-                    <span className="text-stone-800">
-                  {isBuy
-                      ? `${capacity.availableBalance.toLocaleString()}원`
-                      : `${capacity.availableQuantity.toLocaleString()}주`}
-                </span>
-                  </div>
-                  <div className="flex justify-between text-[10px] font-bold">
-                    <span className="text-stone-400">총 주문 금액</span>
-                    <span className="text-stone-800">{numAmount.toLocaleString()}원</span>
-                  </div>
-                  <div className="flex items-center gap-1.5 text-[10px] text-[#a07828] font-bold bg-[#fef6dc] px-3 py-2 rounded-lg">
-                    <MoreHorizontal size={12} />
-                    지정가 주문은 체결 전까지 대기 상태로 유지됩니다
-                  </div>
-                </div>
-
-                {/* 주문 버튼 */}
-                <button
-                    onClick={handleSubmit}
-                    disabled={submitting}
-                    className={cn(
-                        'w-full py-3.5 text-white rounded-md font-black text-sm transition-colors disabled:opacity-50',
-                        isBuy
-                            ? 'bg-brand-red hover:bg-red-700'
-                            : 'bg-brand-blue hover:bg-blue-700'
-                    )}
-                >
-                  {submitting ? '처리 중...' : (isBuy ? '매수하기' : '매도하기')}
-                </button>
-
-                {/* 주문 결과 메시지 */}
-                {orderMsg && (
-                    <p className={cn(
-                        'text-center text-[10px] font-bold',
-                        orderMsg.type === 'success' ? 'text-green-600' : 'text-red-500'
-                    )}>
-                      {orderMsg.text}
-                    </p>
-                )}
-
-                {/* 비로그인 안내 */}
-                {!isLoggedIn && (
-                    <p className="text-center text-[10px] text-stone-400 font-bold">
-                      주문하려면 로그인이 필요합니다
-                    </p>
-                )}
-              </div>
-          )}
-        </div>
-
-        {passwordModal && (
-            <OrderPinPadModal
-                title={
-                  passwordModal.action === 'create'
-                      ? `${isBuy ? '매수' : '매도'} 주문 확인`
-                      : passwordModal.action === 'update'
-                          ? '주문 수정 확인'
-                          : '주문 취소 확인'
-                }
-                description={
-                  passwordModal.action === 'create'
-                      ? `${isBuy ? '매수' : '매도'} 주문 내역을 확인한 뒤 계좌 비밀번호를 입력해 주세요.`
-                      : passwordModal.action === 'update'
-                          ? '수정할 주문 내역을 확인한 뒤 계좌 비밀번호를 입력해 주세요.'
-                          : '취소할 주문 내역을 확인한 뒤 계좌 비밀번호를 입력해 주세요.'
-                }
-                password={passwordModal.action === 'create' ? accountPassword : editAccountPassword}
-                errorMessage={
-                  passwordModal.action === 'create'
-                      ? (orderMsg?.type === 'error' ? orderMsg.text : null)
-                      : (updateMsg?.orderId === passwordModal.orderId && updateMsg.type === 'error' ? updateMsg.text : null)
-                }
-                submitting={submitting}
-                onChange={value => {
-                  if (passwordModal.action === 'create') setAccountPassword(value);
-                  else setEditAccountPassword(value);
-                }}
-                onClose={closePasswordModal}
-                onConfirm={handlePasswordConfirm}
-            />
-        )}
-
-        {(() => {
-          const spec = getConfirmModalSpec();
-          if (!spec) return null;
-          return (
-              <OrderExecutionConfirmModal
-                  title={spec.title}
-                  items={spec.items}
-                  submitting={submitting}
-                  onClose={() => setConfirmModal(null)}
-                  onConfirm={handleConfirmExecution}
-              />
-          );
-        })()}
-      </div>
-  );
-}
